+++ lib/src/in_app_browser/in_app_browser.dart
@@ -377,16 +377,21 @@ class InAppBrowser {
   ///**Official iOS API**: https://developer.apple.com/documentation/uikit/uiscrollviewdelegate/1619392-scrollviewdidscroll
   void onScrollChanged(int x, int y) {}
 
-  ///Event fired when [InAppBrowser] recognizes and starts a downloadable file.
+  ///Use [onDownloadStartRequest] instead
+  @Deprecated('Use `onDownloadStartRequest` instead')
+  void onDownloadStart(Uri url) {}
+
+  ///Event fired when [WebView] recognizes a downloadable file.
+  ///To download the file, you can use the [flutter_downloader](https://pub.dev/packages/flutter_downloader) plugin.
   ///
-  ///[url] represents the url of the file.
+  ///[downloadStartRequest] represents the request of the file to download.
   ///
   ///**NOTE**: In order to be able to listen this event, you need to set [InAppWebViewOptions.useOnDownloadStart] option to `true`.
   ///
   ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebView#setDownloadListener(android.webkit.DownloadListener)
   ///
   ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wknavigationdelegate/1455643-webview
-  void onDownloadStart(Uri url) {}
+  void onDownloadStartRequest(DownloadStartRequest downloadStartRequest) {}
 
   ///Event fired when the [InAppBrowser] webview finds the `custom-scheme` while loading a resource. Here you can handle the url request and return a [CustomSchemeResponse] to load a specific resource encoded to `base64`.
   ///