+++ lib/src/in_app_webview/in_app_webview_controller.dart
@@ -201,15 +201,26 @@ class InAppWebViewController {
             _inAppBrowser!.onScrollChanged(x, y);
         }
         break;
-      case "onDownloadStart":
-        if ((_webview != null && _webview!.onDownloadStart != null) ||
+      case "onDownloadStartRequest":
+        if ((_webview != null &&
+            // ignore: deprecated_member_use_from_same_package
+            (_webview!.onDownloadStart != null || _webview!.onDownloadStartRequest != null)) ||
             _inAppBrowser != null) {
-          String url = call.arguments["url"];
-          Uri uri = Uri.parse(url);
-          if (_webview != null && _webview!.onDownloadStart != null)
-            _webview!.onDownloadStart!(this, uri);
-          else
-            _inAppBrowser!.onDownloadStart(uri);
+          Map<String, dynamic> arguments = call.arguments.cast<String, dynamic>();
+          DownloadStartRequest downloadStartRequest = DownloadStartRequest.fromMap(arguments)!;
+
+          if (_webview != null) {
+            if (_webview!.onDownloadStartRequest != null)
+              _webview!.onDownloadStartRequest!(this, downloadStartRequest);
+            else {
+              // ignore: deprecated_member_use_from_same_package
+              _webview!.onDownloadStart!(this, downloadStartRequest.url);
+            }
+          } else {
+            // ignore: deprecated_member_use_from_same_package
+            _inAppBrowser!.onDownloadStart(downloadStartRequest.url);
+            _inAppBrowser!.onDownloadStartRequest(downloadStartRequest);
+          }
         }
         break;
       case "onLoadResourceCustomScheme":
@@ -375,6 +386,7 @@ class InAppWebViewController {
           } else {
             // ignore: deprecated_member_use_from_same_package
             _inAppBrowser!.androidOnScaleChanged(oldScale, newScale);
+            _inAppBrowser!.onZoomScaleChanged(oldScale, newScale);
           }
         }
         break;